Fix popularity float value

+++ db/migrations/20220402194534_AddPerson.php
@@ -19,7 +19,7 @@ final class AddPerson extends AbstractMigration
                 `id` INT(10) UNSIGNED NOT NULL AUTO_INCREMENT,
                 `name` VARCHAR(256) NOT NULL,
                 `gender` ENUM('0', '1', '2') NOT NULL,
-                `popularity` FLOAT (5,3),
+                `popularity` FLOAT (7,3),
                 `known_for_department` VARCHAR(256),
                 `tmdb_id` INT(10) UNSIGNED DEFAULT NULL,
                 `created_at` TIMESTAMP NOT NULL DEFAULT NOW(),
